app/requirements.txt
matplotlib
numpy
streamlit
tensorflow
Keras
Pillow